Refactor metadata search structure with cherry …
pick Monica’s work
