+++ src/main/java/uk/ac/ebi/atlas/experimentpage/tsneplot/TSnePlotService.java
@@ -93,12 +93,5 @@ public class TSnePlotService {
                                                                 MISSING_METADATA_VALUE_PLACEHOLDER))))
                         .collect(groupingBy(TSnePoint::metadata, mapping(identity(), toImmutableSet()))));
     }
-    public ImmutableMap<String, Map<String, String>>  fetchCellTypeMetadata(String characteristicName,
-                                                             String characteristicValue) {
-        var metadataValuesForCells = cellMetadataDao.getCellTypeMetadata(
-                        characteristicName,
-                        characteristicValue
-                );
-        return ImmutableMap.copyOf(metadataValuesForCells);
-    }
+
 }
